Move new URL query parameter tests to separate class because data is changed in ListingST.
